DailyTasks.java: abstract out file management to a separate class
